Extract out common ModelStub to testutil
